+++ identity/webapp/pages/api/auth/signup.ts
@@ -1,11 +1,12 @@
-import auth0 from '@weco/identity/utils/auth0';
+// TODO: Handle in middleware
+// import { auth0 } from '@weco/identity/lib/auth0';
+// // This will redirect the user directly to the sign-up page.
+// //
+// // See
+// // https://community.auth0.com/t/how-do-i-redirect-users-directly-to-the-hosted-signup-page/42520
 
-// This will redirect the user directly to the sign-up page.
-//
-// See
-// https://community.auth0.com/t/how-do-i-redirect-users-directly-to-the-hosted-signup-page/42520
-// https://github.com/auth0/nextjs-auth0/issues/16#issuecomment-898565337
-export default async (req, res) =>
-  auth0.handleLogin(req, res, {
-    authorizationParams: { screen_hint: 'signup' },
-  });
+// // https://github.com/auth0/nextjs-auth0/issues/16#issuecomment-898565337
+// export default async (req, res) =>
+//   auth0.handleLogin(req, res, {
+//     authorizationParams: { screen_hint: 'signup' },
+//   });
