+++ identity/webapp/pages/index.tsx
@@ -62,7 +62,6 @@ import {
 import { useRequestedItems } from '@weco/identity/hooks/useRequestedItems';
 import { useSendVerificationEmail } from '@weco/identity/hooks/useSendVerificationEmail';
 import { auth0 } from '@weco/identity/lib/auth0';
-import { withPageAuthRequiredSSR } from '@weco/identity/utils/auth0';
 
 type DetailProps = {
   label: string;
@@ -156,70 +155,69 @@ type Props = {
   }; // Key-value store for the user's claims.
 };
 
-export const getServerSideProps: GetServerSideProps<Props | AppErrorProps> =
-  withPageAuthRequiredSSR({
-    getServerSideProps: async context => {
-      const serverData = await getServerData(context);
+export const getServerSideProps: GetServerSideProps<
+  Props | AppErrorProps
+> = async context => {
+  const serverData = await getServerData(context);
 
-      // When a user goes through the registration flow, we create a user
-      // with a placeholder name and *then* update their name in Sierra.
-      //
-      // This causes an issue if they go from the registration flow directly
-      // to their account page:
-      //
-      //    - We can't update their name directly with the Auth0 management
-      //      API, because it's synced from Sierra.
-      //    - Auth0 syncs data from Sierra whenever the user authenticates,
-      //      i.e. logs in [2], and despite trying I haven't found an easy way
-      //      to get Auth0 to re-sync without a login.
-      //
-      // So to get something working for sign-up, we assume that anybody who
-      // signs in with the placeholder surname has just signed up, and we log
-      // them out and redirect to the /success page.
-      //
-      // When they log back in, Auth0 will re-sync their data and pick up their
-      // updated name.
-      //
-      // This is less than ideal, but it gets _something_ working, and we can
-      // revisit re-syncing user data without a login later.
-      //
-      // [1]: https://wellcome.slack.com/archives/CUA669WHH/p1656325929053499?thread_ts=1656322401.443269&cid=CUA669WHH
-      // [2]: https://auth0.com/docs/manage-users/user-accounts/user-profiles#caching-user-profiles
-      //
-      const session = await auth0.getSession();
+  // When a user goes through the registration flow, we create a user
+  // with a placeholder name and *then* update their name in Sierra.
+  //
+  // This causes an issue if they go from the registration flow directly
+  // to their account page:
+  //
+  //    - We can't update their name directly with the Auth0 management
+  //      API, because it's synced from Sierra.
+  //    - Auth0 syncs data from Sierra whenever the user authenticates,
+  //      i.e. logs in [2], and despite trying I haven't found an easy way
+  //      to get Auth0 to re-sync without a login.
+  //
+  // So to get something working for sign-up, we assume that anybody who
+  // signs in with the placeholder surname has just signed up, and we log
+  // them out and redirect to the /success page.
+  //
+  // When they log back in, Auth0 will re-sync their data and pick up their
+  // updated name.
+  //
+  // This is less than ideal, but it gets _something_ working, and we can
+  // revisit re-syncing user data without a login later.
+  //
+  // [1]: https://wellcome.slack.com/archives/CUA669WHH/p1656325929053499?thread_ts=1656322401.443269&cid=CUA669WHH
+  // [2]: https://auth0.com/docs/manage-users/user-accounts/user-profiles#caching-user-profiles
+  //
+  const session = await auth0.getSession();
 
-      if (!session)
-        return {
-          props: serialiseProps({
-            serverData,
-          }),
-        };
+  if (!session)
+    return {
+      props: serialiseProps({
+        serverData,
+      }),
+    };
 
-      if (
-        session.user.email &&
-        session.user.family_name === 'Auth0_Registration_tempLastName'
-      ) {
-        const successParams = new URLSearchParams();
-        successParams.append('email', session.user.email);
+  if (
+    session.user.email &&
+    session.user.family_name === 'Auth0_Registration_tempLastName'
+  ) {
+    const successParams = new URLSearchParams();
+    successParams.append('email', session.user.email);
 
-        const params = new URLSearchParams();
-        params.append('returnTo', `/success?${successParams}`);
+    const params = new URLSearchParams();
+    params.append('returnTo', `/success?${successParams}`);
 
-        return {
-          redirect: {
-            destination: `/api/auth/logout?${params}`,
-            permanent: false,
-          },
-        };
-      }
+    return {
+      redirect: {
+        destination: `/api/auth/logout?${params}`,
+        permanent: false,
+      },
+    };
+  }
 
-      return {
-        props: serialiseProps({
-          serverData,
-        }),
-      };
-    },
-  });
+  return {
+    props: serialiseProps({
+      serverData,
+    }),
+  };
+};
 
 const AccountPage: NextPage<Props> = ({ user: auth0UserClaims }) => {
   const {
