+++ identity/webapp/utils/auth0.ts
@@ -1,46 +1,46 @@
 // TODO ADDRESS THIS PROPERLY https://github.com/wellcomecollection/wellcomecollection.org/issues/10980
 // eslint-disable-next-line
 // @ts-nocheck
-import {
-  // initAuth0,
-  PageRoute,
-  WithPageAuthRequiredPageRouterOptions,
-} from '@auth0/nextjs-auth0';
-import { GetServerSidePropsContext } from 'next';
-import getConfig from 'next/config';
-import { ParsedUrlQuery } from 'querystring';
+// import {
+//   initAuth0,
+//   PageRoute,
+//   WithPageAuthRequiredPageRouterOptions,
+// } from '@auth0/nextjs-auth0';
+// import { GetServerSidePropsContext } from 'next';
+// import getConfig from 'next/config';
+// import { ParsedUrlQuery } from 'querystring';
 
-const { serverRuntimeConfig } = getConfig();
-const config = serverRuntimeConfig as {
-  sessionKeys: string[];
-  sessionVersion: string;
-  siteBaseUrl: string;
-  identityBasePath: string;
-  auth0: {
-    domain: string;
-    clientID: string;
-    clientSecret: string;
-  };
-  remoteApi: {
-    host: string;
-    apiKey: string;
-  };
-};
+// const { serverRuntimeConfig } = getConfig();
+// const config = serverRuntimeConfig as {
+//   sessionKeys: string[];
+//   sessionVersion: string;
+//   siteBaseUrl: string;
+//   identityBasePath: string;
+//   auth0: {
+//     domain: string;
+//     clientID: string;
+//     clientSecret: string;
+//   };
+//   remoteApi: {
+//     host: string;
+//     apiKey: string;
+//   };
+// };
 
-const identityApiScopes = [
-  'create:requests',
-  'delete:patron',
-  'read:user',
-  'read:requests',
-  'send:verification-emails',
-  'update:email',
-  'update:password',
-];
+// const identityApiScopes = [
+//   'create:requests',
+//   'delete:patron',
+//   'read:user',
+//   'read:requests',
+//   'send:verification-emails',
+//   'update:email',
+//   'update:password',
+// ];
 
-const utilityScopes = [
-  'openid', // Required, also gives the token access to the userinfo endpoint
-  'offline_access', // Enables refresh tokens
-];
+// const utilityScopes = [
+//   'openid', // Required, also gives the token access to the userinfo endpoint
+//   'offline_access', // Enables refresh tokens
+// ];
 
 // Things we want in the JWT
 //
@@ -50,17 +50,20 @@ const utilityScopes = [
 // https://openid.net/specs/openid-connect-core-1_0.html#StandardClaims
 // you can actually only ask for quite general *scopes*:
 // https://openid.net/specs/openid-connect-core-1_0.html#ScopeClaims
-const profileScopes = [
-  'profile',
-  'email',
-  'weco:patron_barcode',
-  'weco:patron_role',
-];
+// const profileScopes = [
+//   'profile',
+//   'email',
+//   'weco:patron_barcode',
+//   'weco:patron_role',
+// ];
 
-export const identityAuthorizationParams = {
-  audience: config.remoteApi.host,
-  scope: [...utilityScopes, ...profileScopes, ...identityApiScopes].join(' '),
-};
+// const identityAuthorizationParams = {
+//   audience: {
+//     host: process.env.IDENTITY_API_HOST || 'build',
+//     apiKey: process.env.IDENTITY_API_KEY || 'build',
+//   },
+//   scope: [...utilityScopes, ...profileScopes, ...identityApiScopes].join(' '),
+// };
 
 // const ONE_HOUR_S = 60 * 60;
 // const ONE_DAY_S = 24 * ONE_HOUR_S;
@@ -96,14 +99,14 @@ export const identityAuthorizationParams = {
 
 // Wrapping the built-in method as per https://github.com/auth0/nextjs-auth0#base-path-and-internationalized-routing
 // in order to preserve basePath
-export const withPageAuthRequiredSSR =
-  <P>(
-    opts?: WithPageAuthRequiredPageRouterOptions<P>
-  ): PageRoute<P, ParsedUrlQuery> =>
-  (ctx: GetServerSidePropsContext) =>
-    auth0.withPageAuthRequired<P>({
-      ...opts,
-      returnTo: config.identityBasePath + (opts.returnTo ?? ctx.resolvedUrl),
-    })(ctx);
+// export const withPageAuthRequiredSSR =
+//   <P>(
+//     opts?: WithPageAuthRequiredPageRouterOptions<P>
+//   ): PageRoute<P, ParsedUrlQuery> =>
+//   (ctx: GetServerSidePropsContext) =>
+//     auth0.withPageAuthRequired<P>({
+//       ...opts,
+//       returnTo: config.identityBasePath + (opts.returnTo ?? ctx.resolvedUrl),
+//     })(ctx);
 
-export default auth0;
+// export default auth0;
